[XENFIN-1184] Move VersionCreateDialog out of its subfolder
