+++ __tests__/__image_snapshots__/storyshots-ts-image-storyshots-versions-version-create-dialog-empty-1-snap.png
