--- src/versions/CreateVersion/index.ts
@@ -1,2 +1,0 @@
-export { default as CreateVersionDialog } from "./CreateVersionDialog";
-export { default as CreateVersionFab } from "./CreateVersionFab";
