--- src/versions/CreateVersion/types.ts
@@ -1,10 +1,0 @@
-export enum VersionPanelCreateVersionType {
-    MINOR = "minor",
-    MAJOR = "major",
-}
-
-export interface IVersionPanelCreateVersion {
-    type: VersionPanelCreateVersionType;
-    comment: string;
-    file: File;
-}
