+++ src/versions/types.ts
@@ -1,0 +1,10 @@
+export enum VersionCreateVersionType {
+    MINOR = "minor",
+    MAJOR = "major",
+}
+
+export interface IVersionCreateVersion {
+    type: VersionCreateVersionType;
+    comment: string;
+    file: File;
+}
